Moved the postgres type maps into feast.type_map
