+++ sdk/python/feast/infra/offline_stores/contrib/postgres.py
@@ -31,7 +31,6 @@ from feast.infra.offline_stores.offline_store import (
     RetrievalMetadata,
 )
 from feast.infra.utils.postgres.postgres_config import PostgreSQLConfig
-from feast.infra.utils.postgres.type_map import pg_type_code_to_arrow
 from feast.infra.utils.postgres.utils import (
     _get_conn,
     df_to_postgres_table,
@@ -41,6 +40,7 @@ from feast.on_demand_feature_view import OnDemandFeatureView
 from feast.registry import Registry
 from feast.repo_config import RepoConfig
 from feast.saved_dataset import SavedDatasetStorage
+from feast.type_map import pg_type_code_to_arrow
 
 from .postgres_source import PostgreSQLSource
